PORTFOLIO FONT SIZE & IMG
The image is too large as well as the font size and it pushed everything out of view. Resizing after looking at it on my iphone 6s.

I still don't know why the dev tools shows everything in view on every mobile setting, but on my iphone, it looks different.