ABOUT IMG / FOOTER
Once again, resizing the img on the about page.

This time I added an ID on it to see if I could get it to resize. So far the changes to the width using the selector '.about-photos > img' hasn't worked.

I also moved the entire photos up a little so it doesnt cover the footer, and I moved the footer down just a bit.