ABOUT IMG
Adding the bonus class did remove the image, so I'm removing the bonus form the mobile image to try to resize it once again.

I'm pretty sure I was making the image larger after looking at the standard.css file.

In standard the width is at 15%, so in mobile im going to try 5% to see what happens